+++ R/app_server.R
@@ -35,7 +35,8 @@ app_server <- function(input, output, session) {
     intervention_date = datinput$intervention_date,
     date_ext = datinput$date_ext,
     pad_signals_choice = datinput$pad_signals_choice,
-    min_cases_signals = datinput$min_cases_signals
+    min_cases_signals = datinput$min_cases_signals,
+    filter_outbreak_cases = datinput$filter_outbreak_cases
   )
 
   mod_tabpanel_report_server("report",
@@ -49,7 +50,8 @@ app_server <- function(input, output, session) {
     signals_padded = signals_output$signals_padded,
     signals_agg = signals_output$signals_agg,
     intervention_date = datinput$intervention_date,
-    selected_filter_vars = datinput$selected_filter_vars
+    selected_filter_vars = datinput$selected_filter_vars,
+    filter_outbreak_cases = datinput$filter_outbreak_cases
   )
 
   mod_tabpanel_linelist_server("linelist",