+++ R/get_signals.R
@@ -22,6 +22,11 @@
 #' @param alpha_upper numeric between 0.001 and 0.2 (default: 0.05). Ears and cusum do not use the value; for these, the argument is ignored
 #'   and internally set to NULL.
 #'   Specifies the p-value cutoff used to compute the threshold; for example, a value of 0.05 corresponds to using the 0.95 quantile.
+#' @param filter_outbreak_cases A boolean specifying whether outbreak-associated cases should be excluded from case counts.
+#'   This option can only be applied when GLM-based outbreak detection models are used. For other models, `filter_outbreak_cases = TRUE` is ignored.
+#'   If `preprocessed_data` does not contain an `outbreak_status` column indicating the number of cases associated with outbreaks,
+#'   no filtering is applied, even when `filter_outbreak_cases = TRUE`.
+#'   Default is `FALSE`.
 #'
 #' @return A tibble with columns for signals, expected values, thresholds, and
 #'   stratification information (if applicable), containing both stratified and
@@ -46,7 +51,8 @@ get_signals_all <- function(preprocessed_data,
                             date_ext = NULL,
                             date_var = "date_report",
                             number_of_weeks = 6,
-                            alpha_upper = 0.05) {
+                            alpha_upper = 0.05,
+                            filter_outbreak_cases) {
   results <- get_signals(
     data = preprocessed_data,
     method = method,
@@ -57,7 +63,8 @@ get_signals_all <- function(preprocessed_data,
     date_ext = date_ext,
     date_var = date_var,
     number_of_weeks = number_of_weeks,
-    alpha_upper = alpha_upper
+    alpha_upper = alpha_upper,
+    filter_outbreak_cases = filter_outbreak_cases
   )
   # when stratified signals were computed also add unstratified signals to the dataframe so that all can be visualised
   if (!is.null(stratification)) {
@@ -71,7 +78,8 @@ get_signals_all <- function(preprocessed_data,
       date_ext = date_ext,
       date_var = date_var,
       number_of_weeks = number_of_weeks,
-      alpha_upper = alpha_upper
+      alpha_upper = alpha_upper,
+      filter_outbreak_cases = filter_outbreak_cases
     )
     results <- dplyr::bind_rows(results, results_unstratified)
   }
@@ -99,6 +107,11 @@ get_signals_all <- function(preprocessed_data,
 #' @param alpha_upper numeric between 0.001 and 0.2 (default: 0.05). Ears and cusum do not use the value; for these, the argument is ignored
 #'   and internally set to NULL.
 #'   Specifies the p-value cutoff used to compute the threshold; for example, a value of 0.05 corresponds to using the 0.95 quantile.
+#' @param filter_outbreak_cases A boolean specifying whether outbreak-associated cases should be excluded from case counts.
+#'   This option can only be applied when GLM-based outbreak detection models are used. For other models, `filter_outbreak_cases = TRUE` is ignored.
+#'   If `data` does not contain an `outbreak_status` column indicating the number of cases associated with outbreaks,
+#'   no filtering is applied, even when `filter_outbreak_cases = TRUE`.
+#'   Default is `FALSE`.
 #' @return A tibble containing the results of the signal detection analysis
 #'   stratified by the specified columns.
 #'
@@ -124,7 +137,8 @@ get_signals_stratified <- function(data,
                                    date_ext = NULL,
                                    date_var = "date_report",
                                    number_of_weeks = 6,
-                                   alpha_upper = 0.05) {
+                                   alpha_upper = 0.05,
+                                   filter_outbreak_cases = FALSE) {
   # check that all columns are present in the data
   for (col in stratification_columns) {
     checkmate::assert(
@@ -174,6 +188,18 @@ get_signals_stratified <- function(data,
     checkmate::check_integerish(number_of_weeks)
   )
 
+  if (model != ""){
+  checkmate::assert(
+    checkmate::check_true(filter_outbreak_cases),
+    checkmate::check_false(filter_outbreak_cases),
+    combine = "or"
+  )
+  } else if (model == ""){
+    checkmate::assert(
+      checkmate::check_false(filter_outbreak_cases)
+    )
+    }
+
   # Initialize an empty list to store results per category
   category_results <- list()
 
@@ -212,7 +238,7 @@ get_signals_stratified <- function(data,
       # filter the data
       filter_by_date(date_var = date_var, date_start = date_start, date_end = date_end) %>%
       # aggregate data
-      aggregate_data(date_var = date_var, date_start = date_start, date_end = date_end, date_ext = date_ext, group = category)
+      aggregate_data(date_var = date_var, date_start = date_start, date_end = date_end, date_ext = date_ext, group = category, filter_outbreak_cases = filter_outbreak_cases)
 
     split_list <- sub_data %>%
       dplyr::group_split(!!rlang::sym(category), .keep = FALSE)
@@ -290,10 +316,15 @@ get_signals_stratified <- function(data,
 #' @param date_end A date object or character of format yyyy-mm-dd specifying the end date to filter the data by. Default is NULL.
 #' @param date_ext A date object or character of format yyyy-mm-dd. Extends the aggregated dataset until this date. Default is NULL
 #' @param date_var a character specifying the date variable name used for the aggregation. Default is "date_report".
+#' @param number_of_weeks integer, specifying number of weeks to generate signals for.
 #' @param alpha_upper numeric between 0.001 and 0.2 (default: 0.05). Ears and cusum do not use the value; for these, the argument is ignored
 #'   and internally set to NULL.
 #'   Specifies the p-value cutoff used to compute the threshold; for example, a value of 0.05 corresponds to using the 0.95 quantile.
-#' @param number_of_weeks integer, specifying number of weeks to generate signals for.
+#' @param filter_outbreak_cases A boolean specifying whether outbreak-associated cases should be excluded from case counts.
+#'   This option can only be applied when GLM-based outbreak detection models are used. For other models, `filter_outbreak_cases = TRUE` is ignored.
+#'   If `data` does not contain an `outbreak_status` column indicating the number of cases associated with outbreaks,
+#'   no filtering is applied, even when `filter_outbreak_cases = TRUE`.
+#'   Default is `FALSE`.
 #' @return A tibble containing the results of the signal detection analysis.
 #' @export
 #'
@@ -315,7 +346,8 @@ get_signals <- function(data,
                         date_ext = NULL,
                         date_var = "date_report",
                         number_of_weeks = 6,
-                        alpha_upper = 0.05) {
+                        alpha_upper = 0.05,
+                        filter_outbreak_cases = FALSE) {
   # check that input method and stratification are correct
   checkmate::assert(
     checkmate::check_choice(method, choices = available_algorithms())
@@ -357,6 +389,18 @@ get_signals <- function(data,
   checkmate::assert(
     checkmate::check_integerish(number_of_weeks)
   )
+
+  if (grepl("glm", method)) {
+    checkmate::assert(
+      checkmate::check_true(filter_outbreak_cases),
+      checkmate::check_false(filter_outbreak_cases),
+      combine = "or"
+    )
+  } else {
+    checkmate::assert(
+      checkmate::check_false(filter_outbreak_cases)
+    )
+  }
 
   model <- ""
   time_trend <- FALSE
@@ -405,7 +449,7 @@ get_signals <- function(data,
       # filter the data
       filter_by_date(date_start = date_start, date_end = date_end, date_var = date_var) %>%
       # aggregate and complete the data
-      aggregate_data(date_var = date_var, date_start = date_start, date_end = date_end, date_ext = date_ext)
+      aggregate_data(date_var = date_var, date_start = date_start, date_end = date_end, date_ext = date_ext, filter_outbreak_cases = filter_outbreak_cases)
 
     if (grepl("glm", method)) {
       results <- fun(data_agg, number_of_weeks, model = model, alpha_upper = alpha_upper, time_trend = time_trend, intervention_date = intervention_date)
@@ -431,7 +475,8 @@ get_signals <- function(data,
       date_ext = date_ext,
       date_var = date_var,
       number_of_weeks = number_of_weeks,
-      alpha_upper = alpha_upper
+      alpha_upper = alpha_upper,
+      filter_outbreak_cases = filter_outbreak_cases
     )
   }
 
@@ -568,6 +613,13 @@ pad_signals <- function(data,
   number_of_weeks <- unique(signals$number_of_weeks)
   method <- unique(signals$method)
 
+  # # check if outbreak cases are ignored (could change that as function is not used by glm based algos)
+  # if ("cases_total" %in% names(signals)){
+  #   filter_outbreak_cases <- TRUE
+  # } else{
+  #   filter_outbreak_cases <- FALSE
+  # }
+
   if (grepl("farrington", method)) {
     alpha_upper <- unique(signals$alpha_upper)
   } else {
@@ -591,7 +643,8 @@ pad_signals <- function(data,
       data_no_signals,
       method = method,
       number_of_weeks = timeopt + number_of_weeks,
-      alpha_upper = alpha_upper
+      alpha_upper = alpha_upper,
+      filter_outbreak_cases = FALSE # no filtering used in CUSUM, EARS, FarringtonFlexible
     )
 
     if (!is.null(signals_timeopt)) {
@@ -612,7 +665,8 @@ pad_signals <- function(data,
       date_var = "date_report",
       stratification = stratification,
       number_of_weeks = max_time_opt + number_of_weeks,
-      alpha_upper = alpha_upper
+      alpha_upper = alpha_upper,
+      filter_outbreak_cases = FALSE # no filtering used in CUSUM, EARS, FarringtonFlexible
     ) %>%
       dplyr::select(year, week, upperbound_pad = upperbound, expected_pad = expected, category, stratum) %>%
       dplyr::group_by(category, stratum) %>%